Drop special macOS handling for output streams
we had something somewhat similar in e-antic where we strangely needed
to add an explicit ends() to prevent a segfault. However, str() should
create a copy of the string buffer so there should not be any issues
here. (And I don't understand anymore why the problem was fixed by the
ends in e-antic.)