Bump timeout for sage doctesting
this test takes quite a while on macOS, probably due to all the template
machinery that is involved. It exceeds the 120 seconds timeout in CI:

a = Arb(1)
a += a
a += a
a += a
a += a

Note that later invocations are fast. It's just the initial type
generation that is extremely slow.